Refactored enhanced_connection.rb into two files.
